fix(findRecord): id falsey errors are now on the error property
previously, when id was null or undefined, an error would be thrown, via
dev-time assertion because ember-data does not allow invalid arguments.
However, since findRecord could be used in templates and there is no
try/catch in templates (on purpose), and try/catch is also awkward when
following derived data patterns (mostly that it adds a bit of
verbosity).

Now, instead, when id is null or undefined, a separate error will be
thrown (internally) and the message of that error is stored on the
.error property for access in templates and JS. ember-data is not
invoked when the id is null or undefined.

One additional fix is that ids of 0 are also "truthy", in that id of 0
is passed to ember-data.